feat: Implement tools for retrieving and saving untranslated and translated texts from XLF files
